+++ appstack_plugin/test/appstack_plugin_method_channel_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:appstack_plugin/appstack_plugin_method_channel.dart';
+import 'package:appstack_plugin/src/method_channel_appstack_plugin.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();